# src/main/java/eu/europa/ecb/budgettracker/budgets/dtos/Budget.java
package eu.europa.ecb.budgettracker.budgets.dtos;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.Digits;
import jakarta.validation.constraints.NotNull;

import java.math.BigDecimal;

public record Budget(
        String id,
        @NotNull
        @DecimalMin(value = "0.0", inclusive = false)
        @Digits(integer = 5, fraction = 2)
        BigDecimal amount) {
}